Refactor buckets to nodes
