+++ go/acct/nodes_test.go
@@ -7,13 +7,13 @@ import (
 )
 
 func TestBucketAdd(t *testing.T) {
-	b := NewBuckets()
+	n := NewNodes()
 	now := time.Now()
 	mac, _ := mac.NewFromString("00:22:33:44:55:11")
 	session := SessionID([16]byte{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15})
-	b.Add(1, mac, session, now, 10, 100)
-	b.Add(1, mac, session, now, 10, 100)
-	totals, _ := b.GetBucket(1, mac, session, now)
+	n.Add(1, mac, session, now, 10, 100)
+	n.Add(1, mac, session, now, 10, 100)
+	totals, _ := n.GetBucket(1, mac, session, now)
 	if totals.InBytes != 20 {
 		t.Errorf("InBytes wrong total")
 	}
@@ -21,8 +21,8 @@ func TestBucketAdd(t *testing.T) {
 		t.Errorf("OutBytes wrong total")
 	}
 
-	b.Update(1, mac, session, now, 30, 300)
-	totals, _ = b.GetBucket(1, mac, session, now)
+	n.Update(1, mac, session, now, 30, 300)
+	totals, _ = n.GetBucket(1, mac, session, now)
 	if totals.InBytes != 30 {
 		t.Errorf("InBytes wrong total")
 	}
@@ -32,8 +32,8 @@ func TestBucketAdd(t *testing.T) {
 
 	next := now.Add(1 * time.Second)
 
-	b.Update(1, mac, session, next, 40, 400)
-	totals, _ = b.GetBucket(1, mac, session, next)
+	n.Update(1, mac, session, next, 40, 400)
+	totals, _ = n.GetBucket(1, mac, session, next)
 	if totals.InBytes != 10 {
 		t.Errorf("InBytes wrong total")
 	}
